feat: productionize
