+++ .github/workflows/publish_npm_agentkit_vercel_ai_sdk.yml
@@ -1,0 +1,30 @@
+name: Publish AgentKit Vercel AI SDK to NPM
+
+on:
+  workflow_dispatch:
+
+jobs:
+  deploy-npm-agentkit-vercel-ai-sdk:
+    runs-on: ubuntu-latest
+    permissions:
+      contents: read
+      id-token: write
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-node@v4
+        with:
+          node-version: "18"
+          registry-url: "https://registry.npmjs.org"
+          cache: "npm"
+          cache-dependency-path: ./typescript
+      - name: Install and build
+        working-directory: ./typescript
+        run: |
+          npm ci
+          npm run build
+      - name: Publish @coinbase/agentkit-vercel-ai-sdk
+        working-directory: ./typescript/framework-extensions/vercel-ai-sdk
+        run: |
+          npm publish --ignore-scripts --provenance --access public
+        env:
+          NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
