+++ typescript/.changeset/all-nails-cry.md
@@ -1,0 +1,5 @@
+---
+"@coinbase/agentkit-vercel-ai-sdk": patch
+---
+
+add vercel ai sdk framework support
