+++ typescript/examples/vercel-ai-sdk-cdp-chatbot/.eslintrc.json
@@ -1,4 +1,4 @@
 {
   "parser": "@typescript-eslint/parser",
-  "extends": ["../../../.eslintrc.base.json"]
+  "extends": ["../../.eslintrc.base.json"]
 }
