+++ typescript/examples/vercel-ai-sdk-cdp-chatbot/.prettierignore
@@ -1,0 +1,7 @@
+docs/
+dist/
+coverage/
+.github/
+src/client
+**/**/*.json
+*.md
