+++ typescript/examples/vercel-ai-sdk-cdp-chatbot/README.md
@@ -1,30 +1,64 @@
-# CDP Vercel AI SDK Example - Chatbot
+# CDP AgentKit Vercel AI SDK Extension Examples - Chatbot Typescript
 
-This example demonstrates how to use the CDP Vercel AI SDK to create a chatbot that can interact with the Coinbase Developer Platform.
+This example demonstrates an agent setup as a terminal style chatbot with access to the full set of CDP AgentKit actions.
 
-## Setup
+## Ask the chatbot to engage in the Web3 ecosystem!
 
-1. Install dependencies:
+- "Transfer a portion of your ETH to a random address"
+- "What is the price of BTC?"
+- "Deploy an NFT that will go super viral!"
+- "Deploy an ERC-20 token with total supply 1 billion"
+
+## Prerequisites
+
+### Checking Node Version
+
+Before using the example, ensure that you have the correct version of Node.js installed. The example requires Node.js 18 or higher. You can check your Node version by running:
+
+```bash
+node --version
+```
+
+If you don't have the correct version, you can install it using [nvm](https://github.com/nvm-sh/nvm):
+
+```bash
+nvm install node
+```
+
+This will automatically install and use the latest version of Node.
+
+### API Keys
+
+You'll need the following API keys:
+
+- [CDP API Key](https://portal.cdp.coinbase.com/access/api)
+- [OpenAI API Key](https://platform.openai.com/docs/quickstart#create-and-export-an-api-key)
+
+Once you have them, rename the `.env-local` file to `.env` and make sure you set the API keys to their corresponding environment variables:
+
+- "CDP_API_KEY_NAME"
+- "CDP_API_KEY_PRIVATE_KEY"
+- "OPENAI_API_KEY"
+
+## Running the example
+
+From the root directory, run:
+
 ```bash
 npm install
+npm run build
 ```
 
-2. Set up environment variables:
-```bash
-export CDP_API_KEY_NAME=<your-api-key-name>
-export CDP_API_KEY_PRIVATE_KEY=$'<your-private-key>'
-export OPENAI_API_KEY=<your-openai-api-key>
-export NETWORK_ID=base-sepolia  # Optional: Defaults to base-sepolia
-```
+This will install the dependencies and build the packages locally. The chatbot example uses the local `@coinbase/agentkit-vercel-ai-sdk` and `@coinbase/agentkit` packages. If you make changes to the packages, you can run `npm run build` from root again to rebuild the packages, and your changes will be reflected in the chatbot example.
 
-3. Run the chatbot:
+Now from the `typescript/examples/vercel-ai-sdk-cdp-chatbot` directory, run:
+
 ```bash
 npm start
 ```
 
-## Usage
+Select "1. chat mode" and start telling your Agent to do things onchain!
 
-The chatbot can:
-- Execute CDP operations through natural language
-- Handle streaming responses
-- Use all available CDP tools through the Vercel AI SDK 
+## License
+
+Apache-2.0
