+++ typescript/examples/vercel-ai-sdk-cdp-chatbot/tsconfig.json
@@ -1,9 +1,9 @@
 {
-  "extends": "../../../tsconfig.base.json",
+  "extends": "../../tsconfig.base.json",
   "compilerOptions": {
     "preserveSymlinks": true,
     "outDir": "./dist",
     "rootDir": "."
   },
   "include": ["*.ts"]
-} 
+}
