--- typescript/framework-extensions/vercel-ai-sdk/CHANGELOG.md
@@ -1,9 +1,0 @@
-# CDP AgentKit.js Vercel AI SDK Extension Changelog
-
-## [0.0.1] - 2024-03-XX
-
-### Added
-
-- Initial release of the CDP AgentKit.js Vercel AI SDK Extension.
-- Integration with Vercel AI SDK for streaming responses
-- Basic toolkit setup and configuration
