+++ typescript/framework-extensions/vercel-ai-sdk/README.md
@@ -1,47 +1,70 @@
-# AgentKit Extension - Vercel AI SDK Toolkit
+# Coinbase Agentkit Extension - Vercel AI SDK
 
-Vercel AI SDK extension of AgentKit. Enables agentic workflows to interact with onchain actions.
+This package is an extension used to easily plug [AgentKit](https://docs.cdp.coinbase.com/agentkit/docs/welcome) into [AI SDK by Vercel](https://sdk.vercel.ai/docs/introduction).
 
-## Setup
+## Installation
 
-### Prerequisites
+For a single command to install all necessary dependencies, run:
 
-- [CDP API Key](https://portal.cdp.coinbase.com/access/api)
-- Node.js 18 or higher
+```bash
+npm install @coinbase/agentkit-vercel-ai-sdk @coinbase/agentkit ai @ai-sdk/openai
+```
 
-### Installation
+To break it down, this package is:
 
 ```bash
 npm install @coinbase/agentkit-vercel-ai-sdk
 ```
 
-### Environment Setup
-
-Set the following environment variables:
+This package is used alongside AgentKit and AI SDK, so these will need to be installed as well.
 
 ```bash
-export CDP_API_KEY_NAME=<your-api-key-name>
-export CDP_API_KEY_PRIVATE_KEY=$'<your-private-key>'
-export NETWORK_ID=base-sepolia  # Optional: Defaults to base-sepolia
+npm install @coinbase/agentkit ai
+```
+
+Finally, install the model provider you want to use. For example, to use OpenAI, install the `@ai-sdk/openai` package. See [here](https://sdk.vercel.ai/docs/foundations/providers-and-models#ai-sdk-providers) for a list of supported model providers.
+
+```bash
+npm install @ai-sdk/openai
 ```
 
 ## Usage
 
-### Basic Setup
+The main export of this package is the `getVercelAITools` function. This function takes an AgentKit instance and returns an object containing the tools for the AgentKit agent. This object can then be passed to AI SDK.
+
+Here's a snippet of code that shows how to use the `getVercelAITools` function to get the tools for the AgentKit agent.
+
+###### chatbot.ts
 
 ```typescript
 import { getVercelAITools } from "@coinbase/agentkit-vercel-ai-sdk";
 import { AgentKit } from "@coinbase/agentkit";
+import { generateText } from "ai";
+import { openai } from "@ai-sdk/openai";
 
-// Initialize CDP AgentKit
-const agentkit = await AgentKit.configureWithWallet();
+// Get your Coinbase Developer Platform API key from the Portal: https://portal.cdp.coinbase.com/
+// Or, check out one of the other supported wallet providers: https://github.com/coinbase/agentkit/tree/main/typescript/agentkit
+const agentKit = await AgentKit.from({
+  cdpApiKeyName: process.env.CDP_API_KEY_NAME,
+  cdpApiKeyPrivate: process.env.CDP_API_KEY_PRIVATE_KEY,
+});
 
-// Get available tools
-const tools = await getVercelAITools(agentkit);
+const tools = await getVercelAITools(agentKit);
+
+// There are multiple methods to generate text with AI SDK.
+// See here for more information: https://sdk.vercel.ai/docs/ai-sdk-core/generating-text
+const { text } = await generateText({
+  model: openai("gpt-4o"), // Make sure to have OPENAI_API_KEY set in your environment variables
+  system: "You are an onchain AI assistant with access to a wallet.",
+  prompt: "Print wallet details",
+  tools,
+});
+
+console.log(text);
 ```
 
-The toolkit provides access to all CDP AgentKit tools in a Vercel AI SDK compatible format.
+For a full example, see the [AgentKit AI SDK Chatbot Example](https://github.com/coinbase/agentkit/tree/main/typescript/examples/vercel-ai-sdk-cdp-chatbot).
 
 ## Contributing
 
-See [CONTRIBUTING.md](../../CONTRIBUTING.md) for detailed setup instructions and contribution guidelines. 
+We welcome contributions of all kinds! Please see our [Contributing Guide](https://github.com/coinbase/agentkit/blob/main/CONTRIBUTING.md) for detailed setup instructions and contribution guidelines.
