+++ typescript/framework-extensions/vercel-ai-sdk/tsconfig.json
@@ -1,8 +1,9 @@
 {
-  "extends": "../../../tsconfig.base.json",
+  "extends": "../../tsconfig.base.json",
   "compilerOptions": {
     "outDir": "./dist",
-    "rootDir": "./src"
+    "rootDir": "./src",
+    "module": "Node16"
   },
   "include": ["src/**/*.ts"],
   "exclude": ["src/tests"]
